refactoring again
