+++ commands/todo.js
@@ -1,7 +1,7 @@
 const { google } = require('googleapis');
 const { SlashCommandBuilder } = require('discord.js');
 const { subteamData, excludedTabs } = require('../models/data'); // Colors and thread IDs for each tab, tabs to ignore
-const { formatTasks, calculateDates, authenticate, spreadsheetId, getTabs, buildThisWeekEmbed, buildNextWeekEmbed } = require('../models/sheetsHelper'); 
+const { formatTasks, calculateDates, authenticate, spreadsheetId, getTabs, buildThisWeekEmbed, buildNextWeekEmbed, processSheetData } = require('../models/sheetsHelper'); 
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -31,39 +31,9 @@ module.exports = {
 
             // Iterate over the tabs
             for (const tab of tabs) {
-
-                const range = tab;
-                const threadID = subteamData[tab].thread; // Use threadID from ../models/data.js
-                const color = subteamData[tab].color; // Use color from ../models/data.js
-
-                const sheetValuesResponse = await sheets.spreadsheets.values.get({ spreadsheetId, range });
-                const sheetData = sheetValuesResponse.data.values.slice(1);
-
-                // Get a list of the names of all the columns
-                const columns = sheetValuesResponse.data.values[0];
-
-                // Date calculations
-                const currentDate = new Date();
-                const [thisWeekStartDate, thisWeekEndDate, nextWeekStartDate, nextWeekEndDate] = calculateDates(currentDate);
-
-                // Create an array of task objects, each task object is basically a row of the spreadsheet, keys are column names
-                const sheetRows = sheetData.map(row => {
-                    const taskObject = {};
-                    columns.forEach((columnName, index) => {
-                        taskObject[columnName] = row[index];
-                    });
-                    return taskObject;
-                });
-
-            // Build this week embed
-            const thisWeekFormattedData = await buildThisWeekEmbed(tab, color, columns, sheetRows, thisWeekStartDate, thisWeekEndDate);
-
-            // Build next week embed
-            const nextWeekFormattedData = await buildNextWeekEmbed(tab, color, columns, sheetRows, nextWeekStartDate, nextWeekEndDate);
-
-            // Push to array of response embeds
-            responseEmbeds.push({ tab, formattedData: thisWeekFormattedData });
-            responseEmbeds.push({ tab, formattedData: nextWeekFormattedData });
+                const processedData = await processSheetData(sheets, tab);
+                responseEmbeds.push({ tab, formattedData: processedData.thisWeekFormattedData });
+                responseEmbeds.push({ tab, formattedData: processedData.nextWeekFormattedData });
             }
 
             // Send the embeds to the appropriate thread
