+++ events/sheets-listener.js
@@ -1,7 +1,7 @@
 const { EmbedBuilder } = require('discord.js');
 const { google } = require('googleapis');
 const { subteamData, excludedTabs } = require('../models/data');
-const { formatTasks, calculateDates, authenticate, spreadsheetId, buildThisWeekEmbed, getTabs } = require('../models/sheetsHelper');
+const { formatTasks, calculateDates, authenticate, spreadsheetId, buildThisWeekEmbed, getTabs, processSheetData } = require('../models/sheetsHelper');
 
 module.exports = {
   name: 'interactionCreate',
@@ -55,34 +55,12 @@ module.exports = {
             const tabs = await getTabs(response, name, excludedTabs);
 
             let edited;
-            // Edit the embed
+
             for (const tab of tabs) {
-
-              const range = tab;
-              const threadID = subteamData[tab].thread; // Use threadID from ../models/data.js
-              const color = subteamData[tab].color; // Use color from ../models/data.js
-
-              const sheetValuesResponse = await sheets.spreadsheets.values.get({ spreadsheetId, range });
-              const sheetData = sheetValuesResponse.data.values.slice(1);
-
-              // Get a list of the names of all the columns
-              const columns = sheetValuesResponse.data.values[0];
-
-              // Date calculations
-              const currentDate = new Date();
-              const [thisWeekStartDate, thisWeekEndDate, nextWeekStartDate, nextWeekEndDate] = calculateDates(currentDate);
-
-              // Create an array of task objects, each task object is basically a row of the spreadsheet, keys are column names
-              const sheetRows = sheetData.map(row => {
-                  const taskObject = {};
-                  columns.forEach((columnName, index) => {
-                      taskObject[columnName] = row[index];
-                  });
-                  return taskObject;
-              });
-
+                const processedData = await processSheetData(sheets, tab);
+          
           // Build this week embed
-           edited = await buildThisWeekEmbed(tab, color, columns, sheetRows, thisWeekStartDate, thisWeekEndDate);
+           edited = await processedData.thisWeekFormattedData;
             }
 
             // Replace the existing embed with the new one
